Domain Elemente in eigenen Packages oder in common package eingeteilt
