--- wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/WahlbezirkArt.java
@@ -1,5 +1,0 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain;
-
-public enum WahlbezirkArt {
-    UWB, BWB
-}
