+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/common/WahlbezirkArt.java
@@ -1,0 +1,5 @@
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.common;
+
+public enum WahlbezirkArt {
+    UWB, BWB
+}
