+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/domain/handbuch/Handbuch.java
@@ -1,10 +1,9 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.handbuch;
 
-import jakarta.persistence.Column;
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.common.WahltagIdUndWahlbezirksart;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Lob;
-import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -14,23 +13,19 @@ import lombok.Setter;
 import lombok.ToString;
 
 @Entity
-@Table(name = "Ungueltigews")
 @Getter
 @Setter
-@EqualsAndHashCode
-@ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class UngueltigeWahlscheine {
+@EqualsAndHashCode
+@ToString(onlyExplicitlyIncluded = true)
+public class Handbuch {
 
-    // ========= //
-    // Variables //
-    // ========= //
     @EmbeddedId
+    @ToString.Include
     private WahltagIdUndWahlbezirksart wahltagIdUndWahlbezirksart;
 
-    @Column(name = "ungueltigews")
     @NotNull
     @Lob
-    private byte[] ungueltigeWahlscheine;
+    private byte[] handbuch;
 }
