+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/rest/wahlbezirke/WahlbezirkDTO.java
@@ -1,5 +1,6 @@
 package de.muenchen.oss.wahllokalsystem.basisdatenservice.rest.wahlbezirke;
 
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.rest.common.WahlbezirkArtDTO;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import lombok.Builder;
@@ -8,7 +9,7 @@ import lombok.Builder;
 public record WahlbezirkDTO(@NotNull String wahlbezirkID,
                             @NotNull LocalDate wahltag,
                             @NotNull String nummer,
-                            @NotNull WahlbezirkartDTO wahlbezirkart,
+                            @NotNull WahlbezirkArtDTO wahlbezirkart,
                             @NotNull String wahlnummer,
                             @NotNull String wahlID) {
 }
