--- wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/rest/wahlbezirke/WahlbezirkartDTO.java
@@ -1,5 +1,0 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.rest.wahlbezirke;
-
-public enum WahlbezirkartDTO {
-    UWB, BWB
-}
